fix: use ast to parse lazy types using forward refs to cover some corner cases
Using regexp is simple, but there are a lot of corner cases that it
cannot resolve easily.

This changes the approach to use ast to parse annotations and then
add the resolved type to the global namespace, allowing the `_eval_type`
function to properly handle those corner cases for us.